fix: Avoid worker import in router
